[rdar://104924748] chore: move i18nBanner to App component
